Patch MemcacheClient.Dispose()
A bug was identified in the Dispose method of the MemcacheClient. Each Node
would dipose ONLY the Transports that are in the pool. The transports used
to send a request, or trying to reconnect to a server, would not be disposed.

The solution proposed with this commit is to allow the transports to
check whether the client/node is disposed, via a new interface. If that
is the case, then the transport should dispose itself.

Added unit tests related to the client dispose.

Change-Id: I5ff7b4dcf9ade750452432ba82df1eaa79d31b7d