Suppress execution context flow on Socket.SendAsync()
The socket internally caches the first ExecutionContext it encounters and reuse it.
It leads to inconsistent LogicalCallContext which we use (for exemple to store the the Zipkin trace id).
It doesn't really make sense to have the same execution context reused multiple times for different requests.
Hence we suppress it rather than having an "incorrect" one.

Change-Id: If7e23195a9f8e26c88fb0f07c7b3f8e8dba2b68b