Keep sending request to Couchbase warmup node
A bug from Couchbase flags nodes entering in a cluster as warmup while the
rebalance is running.
This change allow to send queries to those nodes even if they are flagged
as warming up.
The possible drawback is in case of real warmup, instead of have a
synchronous fail to queries to a warmup node we could have slow responses
with status TemporaryFailure.

Change-Id: I38c52499a33aae60eb62cc6f2a432bc092b20499
Jira: BES-9501